Add behavior for easy, medium, and hard enemies
All enemies will initially choose a target on the outside of a sphere
with a radius equal to half the distance between the enemy spawn
location and the player, and with a center halfway between the two.
They will begin by
flying towards this target, after which their behavior is as follows:
- Easy enemies will target and fly to the player's current position.
Once they reach this position, they will retarget the player's current
position and fly towards it again, repeating this behavior endlessly.
- Medium enemies will continually target and fly towards the player's
position.
- Hard enemies will target and fly towards the player's position and
begin firing at a set distance away from the player. Once they get
close enough to the player, they will end their "strafe" and chose a
new target behind them and fly to it. Once they reach this target, they
will strafe the player again. This behavior is repeated endlessly.