Add mini map, fix x2 powerup multiplier
Currently, there are two options for the mini map, MiniMapController and
MiniMapController2. MiniMapController is currently active, but to
switch, deactivate the "Player" child of MiniMap in the game scene and
switch which script is active on MiniMap

The x2 multiplier was adding 2 every time the player picked up a
powerup and setting it back to 1 when the powerup deactivates. Now, the
multiplier is multiplied by 2 every time the powerup is picked up. When
it deactivates, it is divided by 2.